# Conflicts:
#	fluXis/Screens/Edit/EditorMap.cs
#	fluXis/Storyboards/Storyboard.cs

+++ fluXis/Screens/Edit/EditorMap.cs
@@ -403,7 +403,7 @@ public class EditorMap : IVerifyContext
 
     private void throwMissingHandler(ITimedObject obj) => throw new ArgumentException($"Type '{obj.GetType().Name}' does not have a change handler associated with it.");
 
-    public void ApplyOffsetToAll(float offset) => notifiers.ForEach(n => n.ApplyOffset(offset));
+    public void ApplyOffsetToAll(double offset) => notifiers.ForEach(n => n.ApplyOffset(offset));
 
     public void Sort()
     {
@@ -467,7 +467,7 @@ public class EditorMap : IVerifyContext
         void Update(ITimedObject obj);
         void Clear();
 
-        void ApplyOffset(float offset);
+        void ApplyOffset(double offset);
 
         bool Matches(Type type);
     }
@@ -553,7 +553,7 @@ public class EditorMap : IVerifyContext
             OnTypedClear?.Invoke();
         }
 
-        public void ApplyOffset(float offset)
+        public void ApplyOffset(double offset)
         {
             foreach (var obj in list)
             {
