+++ fluXis/Storyboards/Storyboard.cs
@@ -111,7 +111,7 @@ public class Storyboard : EditorMap.IChangeNotifier
         OnClear?.Invoke();
     }
 
-    public void ApplyOffset(float offset) => Elements.ForEach(x =>
+    public void ApplyOffset(double offset) => Elements.ForEach(x =>
     {
         x.StartTime += offset;
         x.EndTime += offset;
